test: added tests for email list & fixed some existing school ff tests
