+++ internal/schoolForm/schFormFiller.go
@@ -55,6 +55,9 @@ const (
 	WAVIER_FORM_MEMBER_BEGIN = 6
 	WAVIER_FORM_MEMBER_CAP   = 17
 
+	// Starting row index of email list sheet
+	EMAIL_LIST_BEGIN = 2
+
 	// Sheet id values
 	CAMPUS_SEC_SHEET_ID = 8
 	EMAIL_SHEET_ID      = 9
@@ -251,7 +254,6 @@ func (ff *FormFiller) FillCommonRecordSheet(e *EventInfo, cL *MinProfile, sIdLis
 	if dur := int(e.EndDate.Sub(e.BeginDate).Hours()/24) + 1; dur < 0 {
 		return errors.New("event duration cannot be negative, " + strconv.Itoa(dur))
 	} else {
-        fmt.Printf("WRITING COST AS: %d\n", 10*dur*len(e.Attendants))
 		ew.setCellValue(s, "I8", 10*dur*len(e.Attendants))
 	}
 	ew.setCellValue(s, "C13", e.Drivers)
@@ -512,9 +514,6 @@ func (ff *FormFiller) FillCampusSecurity(e *EventInfo, cL *MinProfile, sId int) 
 	i := CAMPUS_SEC_MEMBER_BEGIN
 	for _, m := range e.Attendants {
 		if m.UserProfile.IsStudent {
-			if err := DuplicateRowWithStyle(ff.excel, s, i, i+1, 'A', 'I'); err != nil {
-				return err
-			}
 			p := m.UserProfile
 			ew.setCellValue(
 				s,
@@ -586,30 +585,28 @@ func FillAttendance(aL []FullAttendance, ew *errSetCellValue, eFile *excelize.Fi
 func (ff *FormFiller) FillEmailList(e *EventInfo, sId int) error {
 	ew := &errSetCellValue{e: ff.excel}
 	s := ff.excel.GetSheetName(sId)
+	r := EMAIL_LIST_BEGIN
 
 	// Writing rescues' & watchers' emails
-	for i, m := range e.Watchers {
-		if err := DuplicateRowWithStyle(ff.excel, s, i, i+1, 'A', 'B'); err != nil {
-			return err
-		}
-		ew.setCellValue(s, "A"+strconv.Itoa(i+1), m.MinProfile.Name)
-		ew.setCellValue(s, "B"+strconv.Itoa(i+1), m.MinProfile.Email)
+	for _, m := range e.Watchers {
+		ew.setCellValue(s, "A"+strconv.Itoa(r), m.MinProfile.Name)
+		ew.setCellValue(s, "B"+strconv.Itoa(r), m.MinProfile.Email)
+		ew.setCellValue(s, "C"+strconv.Itoa(r), m.Role)
+		r++
 	}
-	for i, m := range e.Rescues {
-		if err := DuplicateRowWithStyle(ff.excel, s, i, i+1, 'A', 'B'); err != nil {
-			return err
-		}
-		ew.setCellValue(s, "A"+strconv.Itoa(i+1), m.MinProfile.Name)
-		ew.setCellValue(s, "B"+strconv.Itoa(i+1), m.MinProfile.Email)
+	for _, m := range e.Rescues {
+		ew.setCellValue(s, "A"+strconv.Itoa(r), m.MinProfile.Name)
+		ew.setCellValue(s, "B"+strconv.Itoa(r), m.MinProfile.Email)
+		ew.setCellValue(s, "C"+strconv.Itoa(r), m.Role)
+		r++
 	}
 
 	// Writing attendants' emails
-	for i, m := range e.Attendants {
-		if err := DuplicateRowWithStyle(ff.excel, s, i, i+1, 'A', 'B'); err != nil {
-			return err
-		}
-		ew.setCellValue(s, "A"+strconv.Itoa(i+1), m.UserProfile.Name)
-		ew.setCellValue(s, "B"+strconv.Itoa(i+1), m.UserProfile.Email)
+	for _, m := range e.Attendants {
+		ew.setCellValue(s, "A"+strconv.Itoa(r), m.UserProfile.Name)
+		ew.setCellValue(s, "B"+strconv.Itoa(r), m.UserProfile.Email)
+		ew.setCellValue(s, "C"+strconv.Itoa(r), m.Role)
+		r++
 	}
 
 	if ew.err != nil {
